4-5-2
fix viewmodel

+++ Controllers/ReplyController.cs
@@ -46,6 +46,10 @@ namespace MOBY_API_Core6.Controllers
                 {
                     return BadRequest(ReturnMessage.Create("Account has been suspended"));
                 }
+                if (uid == -1)
+                {
+                    return BadRequest(ReturnMessage.Create("Account not found"));
+                }
                 if (await ReplyRepository.CreateReply(rep, uid))
                 {
                     return Ok(ReturnMessage.Create("success"));
@@ -70,6 +74,10 @@ namespace MOBY_API_Core6.Controllers
                 {
                     return BadRequest(ReturnMessage.Create("Account has been suspended"));
                 }
+                if (uid == -1)
+                {
+                    return BadRequest(ReturnMessage.Create("Account not found"));
+                }
                 if (await ReplyRepository.UpdateReply(rep, uid))
                 {
                     return Ok(ReturnMessage.Create("success"));
@@ -93,6 +101,10 @@ namespace MOBY_API_Core6.Controllers
                 if (uid == 0)
                 {
                     return BadRequest(ReturnMessage.Create("Account has been suspended"));
+                }
+                if (uid == -1)
+                {
+                    return BadRequest(ReturnMessage.Create("Account not found"));
                 }
                 if (await ReplyRepository.DeleteReply(cmtid, uid))
                 {
